Moving to the better system I had in mind (WIP!)
I really didn't like the system where the profile is directly modified. It was not how I saw this feature, it was not userfriendly at all, just 💩 all the way.

Moved to a new way where variations are enabled when triggered. Then, a new functions allows for getting the profile with the active variations applied to it. We will have to call that everywhere it makes sense.